+++ native-certgen/src/lib.rs
@@ -190,6 +190,9 @@ fn extract_config(env: &mut JNIEnv, config: &JObject) -> Result<CertGenParams> {
     let caller_nonce = get_boolean(env, config, "callerNonce")?;
     let unlocked_device_required = get_boolean(env, config, "unlockedDeviceRequired")?;
     let no_auth_required = get_boolean(env, config, "noAuthRequired")?;
+    let user_auth_type = get_int(env, config, "userAuthType")?;
+    let auth_timeout = get_int(env, config, "authTimeout")?;
+    let user_secure_id = get_long(env, config, "userSecureId")?;
     let uid = get_int(env, config, "uid")?;
     let debug_logging = get_boolean(env, config, "debugLogging")?;
 
@@ -239,6 +242,9 @@ fn extract_config(env: &mut JNIEnv, config: &JObject) -> Result<CertGenParams> {
         caller_nonce,
         unlocked_device_required,
         no_auth_required,
+        user_auth_type,
+        auth_timeout,
+        user_secure_id,
         uid,
         debug_logging,
     })